Issue 5030048: Fix [Issue 3266950] LDAP connector does not save LDAP attributes config over 100
The limit of 100 attributes came from the spring definitions where only 100
attributes were defined and there was evem a test case which checked that
additional attributes would get silently ignored. 

This CL removes the hardcoding of the attributes in connectorInstance.xml and
changes the logic of getting the selected attribute from the UI.

These are the steps:

1. Get all the selected attributes in one json single String 
2. Added a javascript which would append the selected attribute to  the json string and which would remove the attribute if a checkbox is unchecked.
3. Changed the server side logic to parse the String based on the json string. 
4. Added junit test cases to check for large number of attributes, 1 attribute
and 0 attributes.