Fix bug 6513826: Add less verbose logging level for MBean messages. Raised the level of the javax.management package to SEVERE, matching the other core packages (such as javax.net and org.springframework).
The connectors share a common logging.properties with differences only
in whitespace (to be corrected) and connector names. This code review
applies to all of the connectors except SharePoint, whose
logging.properties was one iteration older and is being updated, and
AD groups, which copied the SharePoint version.

Code review: http://codereview.appspot.com/6506115/