Rethrow all NamingExceptions as SnapshotRepositoryRuntimeExceptions, instead of as IllegalStateExceptions or returning empty result maps (b/19613118, b/19044850).
The IllegalStateExceptions terminate the monitor thread, and the empty
result maps tell the diffing package to delete everything in the
previous snapshot. A SnapshotRepositoryRuntimeExceptions, on the other
hand, tells the diffing package to abandon the current pass and try
another pass.

Delete the LdapTransientException class and related code.

Manually tested to make sure that config form errors were reported in
the same way as before, and by conditionally throwing an exception in
LdapHandler.get() to make sure it was handled correctly and retried.

Code review: https://codereview.appspot.com/212880043/